part1/python/requirements.txt
python-dotenv==0.6.4
watson-developer-cloud==0.26.0